Add Ferdinand's training data. Test pad displays favoured signature

+++ TrackPad.cpp
@@ -226,9 +226,9 @@ void TrackPad::clearTrack() {
     track.clear();
 }
 
-void TrackPad::setCurrentTrack(vector<pair<int, int>> &trackVector) {
+void TrackPad::setCurrentTrack(vector<pair<double, double>> &trackVector) {
     track.clear();
-    for (pair<int, int> &v : trackVector) {
+    for (pair<double , double> &v : trackVector) {
         Vector2f V(v.first, v.second);
         track.push_back(V);
     }
