+++ TrackPad.h
@@ -37,7 +37,7 @@ public:
     bool previous();
 
     void setDefaultText(string text);
-    void setCurrentTrack(vector<pair<int, int>> &trackVector);
+    void setCurrentTrack(vector<pair<double, double >> &trackVector);
     void drawDefaultText(Color color = COLOR_TEXT);
     void drawTrackCount(Color color = COLOR_TEXT);
     void drawTrackIndex(Color color = COLOR_TEXT);
